refactor(messenger): emit discord import updates via interval
This changes publishing progress updates of the discord import tool to
happen in an interval instead of synchronously whenever an emission was
done.

The benefit of this is that we reduce the amount of signals being sent
which also improves signal handling in clients as they have to handle
less load.