Work on compression from wav to aac
Unfortunately there is a bug that I don't really have want to spend more time resolving, it is really taking too much time to resolve. Basically  hangs for some encodings params, I don't know why, but it hangs indefinitely and that is a big problem and if I try to interupt the process I get a segmentation violation crash. And recover isn't seeming to work. Also it would be better if we can decode AAC and then re-encode to AAC, so this whole PR is a massive wash.