feat: start media server with no db right after logout
In prev PR #3500, I added the function to start media server before login in
`OpenAccounts` This works when app starts up for the first time, since
there's no multiaccounts data and `OpenAccounts` has to be called.

When user logout, mobile code calls logout and get multiaccounts at the
same time, which is fine since we want to get these data as early as
possible to render the view

https://github.com/status-im/status-mobile/blob/eea55b22d73dcb0a5b876c63f24fa9635c42185f/src/status_im/multiaccounts/logout/core.cljs#L17-L25

trigger logout
```clojure
::logout
```

get multiaccounts data with `OpenAccounts`
```clojure
:setup/open-multiaccounts             #(re-frame/dispatch [:setup/initialize-multiaccounts
                                                                          % {:logout? logout?}])
```

After which the media server will be shutdown by `Logout`.

This PR changes the code to start the media server right after logout,
so that we can server user avatars as soon as possible. (It's required
in the view right after user press logout)

Signed-off-by: yqrashawn <namy.19@gmail.com>