Send requests to join to community admins
Requests to join communities are encrypted by the sender such that only
the owner account is able decrypt and process the message.

In order to allow community admins to accept and reject requests to
join, they need to receive the requests as well.

The easy solution is to send multiple messages with the same request to
all community admins.